# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Academic Path Manager API",
    "description": "API for managing academic paths in Computer Science and Social Sciences with OAuth authentication",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://academic-path-manager.onrender.com",
      "description": "Production server (Render)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Student": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "studentId": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "field": {
            "type": "string",
            "enum": ["computer_science", "social_sciences"]
          },
          "enrolledCourses": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "courseCode": { "type": "string" },
                "courseTitle": { "type": "string" },
                "grade": {
                  "type": "string",
                  "enum": ["A", "B", "C", "D", "F", null]
                }
              }
            }
          },
          "totalCredits": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "StudentInput": {
        "type": "object",
        "required": ["studentId", "firstName", "lastName", "email", "field"],
        "properties": {
          "studentId": {
            "type": "string",
            "description": "Student ID",
            "minLength": 3
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "field": {
            "type": "string",
            "enum": ["computer_science", "social_sciences"],
            "description": "Academic field"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/students": {
      "get": {
        "summary": "Get all students",
        "description": "Returns all students, optionally filtered by field (Protected: Authenticated users only)",
        "tags": ["Students"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "description": "Filter by field",
            "schema": {
              "type": "string",
              "enum": ["computer_science", "social_sciences"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Student"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new student",
        "description": "Creates a new student (Protected: Professor or Admin only)",
        "tags": ["Students"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - professor or admin privileges required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "summary": "Get student by ID",
        "description": "Get a specific student by ID (Protected: Authenticated users only)",
        "tags": ["Students"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Student ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Student found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a student",
        "description": "Update an existing student (Protected: Professor or Admin only)",
        "tags": ["Students"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Student ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - professor or admin privileges required"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a student",
        "description": "Delete a student (Protected: Admin only)",
        "tags": ["Students"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Student ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - admin privileges required"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/students/{id}/enroll": {
      "put": {
        "summary": "Enroll student in a course",
        "description": "Enroll a student in a course (Protected: Authenticated users only)",
        "tags": ["Students"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Student ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseCode": {
                    "type": "string",
                    "description": "Course code to enroll in"
                  },
                  "courseTitle": {
                    "type": "string",
                    "description": "Course title"
                  }
                },
                "required": ["courseCode", "courseTitle"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student enrolled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or already enrolled"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}